pyproject.toml
[tool.poetry]
name = "nornir_rich"
version = "0.1.0"
description = "Nornir results printer using "
authors = ["Patrick Avery <patrickdaj@gmail.com>"]

[tool.poetry.scripts]
nornir-cli = "nornir_rich.plugins.cli.nornir_cli:main"

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
nornir = {version = "^3.0.0-beta.1", allow-prereleases = true}
rich = "^6.0.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.3"
black = { version = "19.10b0", allow-prereleases = true }
mypy = "*"
pylama = "^7.7.1"
pytest = "*"
nbval = "*"
jupyter = "^1"
sphinx = "^3.0.3"
sphinxcontrib-napoleon = "^0.7"
nbsphinx = "^0.6.1"
sphinx-issues = "^1.2.0"
sphinx_rtd_theme = "^0.4.3"
nornir_napalm = "^0.1.1"
napalm-vyos = "^0.2.0"
python-dotenv = "^0.14.0"
nornir_utils = { path = '../nornir_utils' }
nornir_netmiko = { path = '../nornir_netmiko' }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
